+++ RELEASE_NOTES.md
@@ -1,10 +1,12 @@
 ### 0.8.4
+* [EventStore][New] EventJournal support reader streams.
 * [WAZ.Storage][New] CloudQueue getting messages operation accept visibility timeout parameter.
 * [WAZ.Storage][New] CloudQueue getting messages operation accept message count parameter.
 * [WAZ.Storage][New] CloudTable expose and accept table continuation token.
 * [EventStore][Improvment] Consumer can start reading from the end of streams.
 * [EventStore][Improvment] Retry notification delivery timeout.
 * [EventStore][Breaking] Notification listener starts reading stream from the end.
+* [EventStore][Breaking] EventJournal.ReadStreamReaderPositionAsync throw exception, if reader is not registered.
 * [EventStore][BugFix] Consumer initialize reader only in leader state.
 
 ### 0.8.3